refactor: move classes
Move the test and extension classes to right package
