+++ engine/src/test/java/io/camunda/zeebe/streamprocessor/NewStreamProcessorReplayModeTest.java
@@ -17,8 +17,6 @@ import static org.mockito.Mockito.timeout;
 
 import io.camunda.zeebe.engine.api.RecordProcessor;
 import io.camunda.zeebe.engine.util.Records;
-import io.camunda.zeebe.engine.util.StreamPlatform;
-import io.camunda.zeebe.engine.util.StreamPlatformExtension;
 import io.camunda.zeebe.protocol.impl.record.value.processinstance.ProcessInstanceRecord;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
